+++ src/FreeElephants/Validation/ValidatorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace FreeElephants\Validation;
+
+interface ValidatorInterface
+{
+    public function validate(array $data, Rules $rules, string $language = null): ValidationResult;
+}