+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Data Transfer Object classes generation from swagger spec
 
+## [0.0.5] - 2020-05-08
+### Added
+- Validation subpackage based on rakit/validation
+- Request Validation Middleware
+- JsonApiResponseFactory
+- I18n support
+
+### Removed
+- Unused free-elephants/php-di dependency (but suggested) 
+
 ## [0.0.4] - 2020-04-27
 ### Fixed 
 - Bump version with merged 0.0.2 changes
@@ -29,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - FastRoute Dispatcher generation from swagger operationIds
 
-[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.4...HEAD
+[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.5...HEAD
+[0.0.5]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.4...0.0.5
 [0.0.4]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.1...0.0.2